+++ backend/src/middleware/alterPeriodPayload.ts
@@ -1,6 +1,10 @@
+/**
+ *Salamaik API 
+ * © Kanwil DJPb Sumbar 2024
+ */
+
 import { Request, Response, NextFunction } from "express";
-import ErrorDetail from "../model/error.model";
-import {superAdmin, adminKanwil, userKanwil} from '../config/role';
+import * as roles from '../config/role';
 
 export default function alterPeriodPayload(req: Request, res: Response, next: NextFunction) {
   const postedPeriod = req.params.period ?? '';
@@ -11,7 +15,7 @@ export default function alterPeriodPayload(req: Request, res: Response, next: Ne
 
   const role = req.payload.role;
 
-  if(role === superAdmin || role === adminKanwil || role === userKanwil){
+  if(role === roles.superAdmin || role === roles.adminKanwil || role === roles.userKanwil){
     req.payload.period = Number(postedPeriod);
   };